+++ src/App.tsx
@@ -3,7 +3,6 @@ import { Routes, Route, Navigate } from 'react-router-dom';
 import { HomePage } from './pages/HomePage';
 import { AboutPage } from './pages/AboutPage';
 import { FormPage } from './pages/FormPage';
-import { FormPageFn } from './pages/FormPage/FromPageFn';
 import { NotFoundPage } from './pages/NotFoundPage';
 
 import './App.scss';
@@ -16,7 +15,6 @@ function App() {
           <Route path="/" element={<HomePage title="Home" />} />
           <Route path="/about" element={<AboutPage title="About" />} />
           <Route path="/form" element={<FormPage title="Form" />} />
-          <Route path="/formFn" element={<FormPageFn title="FormFn" />} />
           <Route path="*" element={<Navigate to="/404" />} />
           <Route path="/404" element={<NotFoundPage title="NotFound" />} />
         </Routes>
